introduce ProxyForwarder
