+++ engine_store_ffi/src/core/mod.rs
@@ -1,0 +1,10 @@
+// Copyright 2022 TiKV Project Authors. Licensed under Apache-2.0.
+
+pub(crate) mod common;
+pub mod fast_add_peer;
+pub mod forward_raft;
+pub mod forwarder;
+
+pub use fast_add_peer::*;
+pub use forward_raft::*;
+pub use forwarder::*;
