+++ new-mock-engine-store/src/server.rs
@@ -14,7 +14,7 @@ use collections::{HashMap, HashSet};
 use concurrency_manager::ConcurrencyManager;
 use encryption_export::DataKeyManager;
 use engine_rocks::RocksSnapshot;
-use engine_store_ffi::observer::DebugStruct;
+use engine_store_ffi::core::DebugStruct;
 use engine_traits::{Engines, MiscExt};
 use futures::executor::block_on;
 use grpcio::{ChannelBuilder, EnvBuilder, Environment, Error as GrpcError, Service};
@@ -537,7 +537,7 @@ impl ServerCluster {
         let max_grpc_thread_count = cfg.server.grpc_concurrency;
         let server_cfg = Arc::new(VersionTrack::new(cfg.server.clone()));
 
-        let packed_envs = engine_store_ffi::observer::PackedEnvs {
+        let packed_envs = engine_store_ffi::core::PackedEnvs {
             engine_store_cfg: cfg.proxy_cfg.engine_store.clone(),
             pd_endpoints: cfg.pd.endpoints.clone(),
         };
